updated  api response test and fixed my folder structure
