+++ app/tests/test_stats.py
@@ -1,4 +1,4 @@
-from app.modules.jobs.service import calculate_job_stats
+from app.modules.stats.service import calculate_job_stats
 
 def test_calculate_job_stats():
     # 1. SETUP: Create a controlled sample of data
@@ -32,6 +32,20 @@ def test_calculate_job_stats():
     # 3. ASSERT: Verify the function returns the exact numbers it should
     assert result["Total_jobs"] == 3
     assert result["Remote_jobs"] == 2
-    
-    # TechFlow shows up twice, Health Analytics shows up once
     assert result["top_companies"] == ["TechFlow", "Health Analytics"]
+
+
+def test_calculate_job_stats_empty_response():
+    # 1. SETUP: Simulate Adzuna finding zero jobs
+    mock_empty_data = {
+        "count": 0,
+        "results": []
+    }
+
+    # 2. EXECUTE: Pass the empty data into your function
+    result = calculate_job_stats(mock_empty_data)
+
+    # 3. ASSERT: Verify the math gracefully handles the zeros
+    assert result["Total_jobs"] == 0
+    assert result["Remote_jobs"] == 0
+    assert result["top_companies"] == []
